iso-routing: Don't call serverRender until final next()
Previously, if the combined middleware chain involved waiting
for an async response, the call to serverRender would happen
before the response arrived.

With this change, the call to serverRender will only occur if
and when the last next() in the middleware chain is called. This
is now in keeping with the app.get() contract and allows async
behaviour in requests.