Themes: Use a TTL in the theme-details server cache
Previous strategy was to call API on every server request and
update the cache if the result was newer than existing. The
complication with this is that in the async part, we don't know
whether or not to continue serving a response (depends
on whether the cache was empty).

Simpler to add a timestamp to the cache and use the cached details
if they are less than an hour old, otherwise wait for the API
response.